chore(video-upload): refatorado testes de Video
- separado testes unitarios de Crud e Upload da classe video;
- separado testes de integracao de Crud e Upload da classe VideoController;
- alterado Dockerfile de desenvolvimento para ter o modulo GD para processamento de imagens;
- criado teste de rollback de arquivos no cadastro do Video

 Changes to be committed:
	modified:   Dockerfile
	modified:   app/Models/Video.php
	modified:   database/migrations/2019_10_10_020756_create_videos_table.php
	new file:   tests/Feature/Http/Controllers/Api/VideoController/BaseVideoControllerTest.php
	renamed:    tests/Feature/Http/Controllers/Api/VideoControllerTest.php -> tests/Feature/Http/Controllers/Api/VideoController/VideoControllerCrudTest.php
	new file:   tests/Feature/Http/Controllers/Api/VideoController/VideoControllerUploadTest.php
	new file:   tests/Feature/Models/Video/BaseVideoTest.php
	renamed:    tests/Feature/Models/VideoTest.php -> tests/Feature/Models/Video/VideoCrudTest.php
	new file:   tests/Feature/Models/Video/VideoUploadTest.php
	new file:   tests/Traits/TestUploads.php
