+++ Dockerfile
@@ -3,8 +3,13 @@ FROM php:7.3.6-fpm-alpine3.9
 LABEL maintainer="Pablo Sousa <pablosousa.ads@gmail.com>"
 
 # install extensions and aplications linux
-RUN apk add --no-cache openssl bash mysql-client nodejs npm
+RUN apk add --no-cache openssl
+RUN apk add bash mysql-client nodejs npm freetype-dev libjpeg-turbo-dev libpng-dev
 RUN docker-php-ext-install pdo pdo_mysql
+
+#extensao php install
+RUN docker-php-ext-configure gd --with-gd --with-freetype-dir=/usr/include/ --with-jpeg-dir=/usr/include --with-png-dir=/usr/include
+RUN docker-php-ext-install -j$(nproc) gd
 
 # install composer
 RUN curl -sS https://getcomposer.org/installer | php -- \
