+++ tests/Feature/Http/Controllers/Api/VideoController/BaseVideoControllerTest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Tests\Feature\Http\Controllers\Api\VideoController;
+
+use App\Models\Video;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
+use Tests\TestCase;
+
+abstract class BaseVideoControllerTest extends TestCase
+{
+    use DatabaseMigrations;
+
+    protected $video;
+    protected $sendData;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->video = factory(Video::class)->create(['opened' => false]);
+
+        $this->sendData = [
+            'title' => 'title',
+            'description' => 'description',
+            'year_launched' => 2010,
+            'rating' => Video::RATING_FREE,
+            'duration' => 10,
+        ];
+    }
+
+    protected function model()
+    {
+        return Video::class;
+    }
+
+    protected function routeStore()
+    {
+        return route('videos.store');
+    }
+
+    protected function routeUpdate()
+    {
+        return route('videos.update', ['video' => $this->video->id]);
+    }
+}
