+++ tests/Feature/Models/Video/BaseVideoTest.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tests\Feature\Models\Video;
+
+use App\Models\Video;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
+use Tests\TestCase;
+
+abstract class BaseVideoTest extends TestCase
+{
+    use DatabaseMigrations;
+
+    protected $data;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->data = [
+            'title' => 'title',
+            'description' => 'description',
+            'year_launched' => 2010,
+            'rating' => Video::RATING_FREE,
+            'duration' => 10,
+        ];
+    }
+}
